Remove ChunkUpdateEvent
