+++ crates/mc-rs-core/src/world/structure/chunk.rs
@@ -11,12 +11,8 @@ use mc_rs_protocol::{
 };
 use parking_lot::RwLock;
 
-use crate::{
-    blocks::BlockData,
-    world::{
-        palette::GlobalPalette, ChunkUpdateEvent, WorldType, Worlds, CHUNK_SIZE,
-        CHUNK_VERT_DISPLACEMENT, SECTION_COUNT,
-    },
+use crate::world::{
+    palette::GlobalPalette, WorldType, CHUNK_SIZE, CHUNK_VERT_DISPLACEMENT, SECTION_COUNT,
 };
 
 use super::section::Section;
@@ -95,105 +91,6 @@ impl Chunk {
         let world = root.get("WORLD_SURFACE")?.as_long_array()?;
 
         Some((motion.to_owned(), world.to_owned()))
-    }
-
-    /// Regenerate chunk meshes.
-    pub fn update_chunk(
-        query: Query<Ref<Chunk>>,
-        block_data: Res<BlockData>,
-        mut worlds: ResMut<Worlds>,
-        mut events: EventWriter<ChunkUpdateEvent>,
-        mut commands: Commands,
-    ) {
-        for chunk in query.iter() {
-            if chunk.is_changed() {
-                let Some(world) = worlds.get_world_mut(&chunk.world_type) else {
-                    error!(
-                        "Failed to get world {:?} with chunk {:?}!",
-                        chunk.world_type, chunk.position
-                    );
-                    continue;
-                };
-
-                // If the chunk is empty, despawn it
-                if chunk.block_count() == 0 {
-                    if let Some(entity) = world.get_chunk_id(&chunk.position) {
-                        let entity = **entity;
-                        commands.entity(entity).despawn_recursive();
-                        world.remove(&chunk.position);
-                    } else {
-                        error!(
-                            "Failed to get chunk entity for empty chunk {:?}",
-                            chunk.position
-                        );
-                    }
-                    continue;
-                }
-
-                let mut neighbors = [None, None, None, None];
-                for (pos, val) in chunk
-                    .position
-                    .around()
-                    .into_iter()
-                    .zip(neighbors.iter_mut())
-                {
-                    if let Some(chunk) = world.get_chunk_ref(&query, &pos) {
-                        if chunk.block_count() != 0 {
-                            *val = Some(chunk.sections.clone());
-                        }
-                    }
-                }
-
-                if let Some(entity) = world.get_chunk_id(&chunk.position) {
-                    events.send(ChunkUpdateEvent::new(
-                        **entity,
-                        chunk.sections.clone(),
-                        neighbors,
-                        block_data.clone(),
-                    ));
-                } else {
-                    error!(
-                        "Failed to get chunk entity for changed chunk {:?}",
-                        chunk.position
-                    );
-                }
-            }
-
-            if chunk.is_added() {
-                // For each neighboring chunk
-                let world = worlds.get_world(&chunk.world_type).unwrap();
-                for neighbor_pos in chunk.position.around().into_iter() {
-                    if let Some(neighbor_chunk) = world.get_chunk_ref(&query, &neighbor_pos) {
-                        // Skip chunks with no blocks
-                        if neighbor_chunk.block_count() == 0 {
-                            continue;
-                        }
-
-                        // Get that chunk's neighbors
-                        let mut neighbors = [None, None, None, None];
-                        for (pos, val) in
-                            neighbor_pos.around().into_iter().zip(neighbors.iter_mut())
-                        {
-                            if let Some(chunk) = world.get_chunk_ref(&query, &pos) {
-                                *val = Some(chunk.sections.clone());
-                            }
-                        }
-
-                        // Update the neighbor chunk mesh
-                        if let Some(entity) = world.get_chunk_id(&neighbor_pos) {
-                            events.send(ChunkUpdateEvent::new(
-                                **entity,
-                                chunk.sections.clone(),
-                                neighbors,
-                                block_data.clone(),
-                            ));
-                        } else {
-                            error!("Failed to get chunk entity for added chunk {neighbor_pos:?}");
-                        }
-                    }
-                }
-            }
-        }
     }
 
     /// Insert data into a chunk section.
